lib/worker: Avoid issuing 'kill(0, SIGKILL)'
It appears we can sometime get child processes with pid 0 when
checking if any worker jobs have timed out. That's pretty bad and
it's still not known why it happens, but when it does we must take
care not to issue kill(0, SIGKILL), since that kills the worker
and all its running jobs, finally leaving Nagios with no workers
and unable to do anything.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2331 5f96b256-904b-4d8d-8c98-d829582c6739

+++ lib/worker.c
@@ -323,6 +323,11 @@ static void kill_job(child_process *cp, int reason)
 	int ret;
 	struct rusage ru;
 
+	if (!cp->ei->pid) {
+		wlog("No pid for job %d (%u running); '%s'", cp->id, running_jobs, cp->cmd);
+		return;
+	}
+
 	/* brutal but efficient */
 	ret = kill(cp->ei->pid, SIGKILL);
 	if (ret < 0) {
